move cdp related logic from debug addpater to cdpSession
